+++ packages/components/card/index.ts
@@ -1,0 +1,6 @@
+import Card from './Card.vue'
+import type { CardProps } from './types'
+
+export { Card }
+export type { CardProps }
+export default Card