+++ packages/components/tag/index.ts
@@ -1,0 +1,6 @@
+import Tag from './Tag.vue'
+import type { TagProps, TagEmits } from './types'
+
+export { Tag }
+export type { TagProps, TagEmits }
+export default Tag